Renamed YoutubeReader module file.
